requirements.txt
langchain-core==0.1.27
langchain-openai==0.0.5
python-dotenv==1.0.0 
